fix: NPE and refactor
- Fix NPE: call parent frame in initalization
- move Handler common method to ProjectUICommands
