+++ src/org/omegat/gui/main/MainWindowMenuHandler.java
@@ -49,25 +49,18 @@ import java.util.regex.Pattern;
 import javax.swing.Action;
 import javax.swing.JCheckBoxMenuItem;
 import javax.swing.JOptionPane;
-import javax.swing.SwingWorker;
 import javax.swing.UIManager;
-import javax.swing.text.JTextComponent;
 
 import org.openide.awt.AbstractMnemonicsAction;
 
 import org.omegat.Main;
 import org.omegat.core.Core;
-import org.omegat.core.CoreEvents;
-import org.omegat.core.KnownException;
-import org.omegat.core.data.ProjectFactory;
 import org.omegat.core.data.SourceTextEntry;
 import org.omegat.core.data.TMXEntry;
 import org.omegat.core.matching.NearString;
 import org.omegat.core.matching.NearString.MATCH_SOURCE;
 import org.omegat.core.search.SearchMode;
-import org.omegat.core.spellchecker.ISpellChecker;
 import org.omegat.core.tagvalidation.ErrorReport;
-import org.omegat.filters2.master.PluginUtils;
 import org.omegat.gui.dialogs.AboutDialog;
 import org.omegat.gui.dialogs.GoToSegmentDialog;
 import org.omegat.gui.dialogs.LastChangesDialog;
@@ -706,7 +699,7 @@ public final class MainWindowMenuHandler {
         @Override
         public void actionPerformed(final ActionEvent e) {
             Log.logInfoRB("LOG_MENU_CLICK", e.getActionCommand());
-            prepareForExit(() -> System.exit(0));
+            ProjectUICommands.prepareForExit(() -> System.exit(0));
         }
     }
 
@@ -726,14 +719,10 @@ public final class MainWindowMenuHandler {
             String projectDir = Core.getProject().isProjectLoaded()
                     ? Core.getProject().getProjectProperties().getProjectRoot()
                     : null;
-            prepareForExit(() -> {
+            ProjectUICommands.prepareForExit(() -> {
                 Main.restartGUI(projectDir);
             });
         }
-    }
-
-    protected static void projectExitAction() {
-        prepareForExit(() -> System.exit(0));
     }
 
     private static void openFile(File path) {
@@ -753,72 +742,6 @@ public final class MainWindowMenuHandler {
             Log.logErrorRB(ex, "RPF_ERROR");
             Core.getMainWindow().displayErrorRB(ex, "RPF_ERROR");
         }
-    }
-
-    private static void prepareForExit(Runnable onCompletion) {
-        // Bug #902: commit the current entry first
-        // We do it before checking project status, so that it can eventually
-        // change it
-        if (Core.getProject().isProjectLoaded()) {
-            Core.getEditor().commitAndLeave();
-        }
-
-        boolean projectModified = false;
-        if (Core.getProject().isProjectLoaded()) {
-            projectModified = Core.getProject().isProjectModified();
-        }
-        // RFE 1302358
-        // Add Yes/No Warning before OmegaT quits
-        if (projectModified || Preferences.isPreference(Preferences.ALWAYS_CONFIRM_QUIT)) {
-            if (JOptionPane.YES_OPTION != JOptionPane.showConfirmDialog(
-                    Core.getMainWindow().getApplicationFrame(), OStrings.getString("MW_QUIT_CONFIRM"),
-                    OStrings.getString("CONFIRM_DIALOG_TITLE"), JOptionPane.YES_NO_OPTION)) {
-                return;
-            }
-        }
-
-        SegmentExportImport.flushExportedSegments();
-
-        new SwingWorker<Void, Void>() {
-            @Override
-            protected Void doInBackground() throws Exception {
-                if (Core.getProject().isProjectLoaded()) {
-                    // Save the list of learned and ignore words
-                    ISpellChecker sc = Core.getSpellChecker();
-                    sc.saveWordLists();
-                    try {
-                        Core.executeExclusively(true, () -> {
-                            Core.getProject().saveProject(true);
-                            ProjectFactory.closeProject();
-                        });
-                    } catch (KnownException ex) {
-                        // hide exception on shutdown
-                    }
-                }
-
-                CoreEvents.fireApplicationShutdown();
-
-                PluginUtils.unloadPlugins();
-
-                return null;
-            }
-
-            @Override
-            protected void done() {
-                try {
-                    get();
-
-                    MainWindowUI.saveScreenLayout((MainWindow) Core.getMainWindow());
-
-                    Preferences.save();
-
-                    onCompletion.run();
-                } catch (Exception ex) {
-                    Log.logErrorRB(ex, "PP_ERROR_UNABLE_TO_READ_PROJECT_FILE");
-                    Core.getMainWindow().displayErrorRB(ex, "PP_ERROR_UNABLE_TO_READ_PROJECT_FILE");
-                }
-            }
-        }.execute();
     }
 
     @SuppressWarnings("serial")
@@ -1085,17 +1008,7 @@ public final class MainWindowMenuHandler {
             if (!Core.getProject().isProjectLoaded()) {
                 return;
             }
-            MainWindowUI.createSearchWindow(SearchMode.SEARCH, getTrimmedSelectedTextInMainWindow());
-        }
-    }
-
-    public static void findInProjectReuseLastWindow() {
-        if (!Core.getProject().isProjectLoaded()) {
-            return;
-        }
-        String text = getTrimmedSelectedTextInMainWindow();
-        if (!MainWindowUI.reuseSearchWindow(text)) {
-            MainWindowUI.createSearchWindow(SearchMode.SEARCH, text);
+            MainWindowUI.createSearchWindow(SearchMode.SEARCH, ProjectUICommands.getTrimmedSelectedTextInMainWindow());
         }
     }
 
@@ -1115,21 +1028,8 @@ public final class MainWindowMenuHandler {
             if (!Core.getProject().isProjectLoaded()) {
                 return;
             }
-            MainWindowUI.createSearchWindow(SearchMode.REPLACE, getTrimmedSelectedTextInMainWindow());
+            MainWindowUI.createSearchWindow(SearchMode.REPLACE, ProjectUICommands.getTrimmedSelectedTextInMainWindow());
         }
-    }
-
-    private static String getTrimmedSelectedTextInMainWindow() {
-        String selection = null;
-        Component component = Core.getMainWindow().getApplicationFrame().getMostRecentFocusOwner();
-        if (component instanceof JTextComponent) {
-            selection = ((JTextComponent) component).getSelectedText();
-            if (!StringUtil.isEmpty(selection)) {
-                selection = EditorUtils.removeDirectionChars(selection);
-                selection = selection.trim();
-            }
-        }
-        return selection;
     }
 
     /** Set active match to #1. */
